+++ internal/ioutil/ioutil_unix.go
@@ -1,5 +1,5 @@
-//go:build windows
-// +build windows
+//go:build !windows
+// +build !windows
 
 /*
 Copyright The ORAS Authors.
@@ -16,8 +16,36 @@ See the License for the specific language governing permissions and
 limitations under the License.
 */
 
-package ioutils
+package ioutil
 
+import (
+	"os"
+	"syscall"
+)
+
+// CopyFilePermissions copies file ownership and permissions from "src" to "dst",
+// ignoring any error during the process.
 func CopyFilePermissions(src, dst string) {
-	// TODO implement for Windows
+	var (
+		mode     os.FileMode = 0600
+		uid, gid int
+	)
+
+	fi, err := os.Stat(src)
+	if err != nil {
+		return
+	}
+	if fi.Mode().IsRegular() {
+		mode = fi.Mode()
+	}
+	if err := os.Chmod(dst, mode); err != nil {
+		return
+	}
+
+	uid = int(fi.Sys().(*syscall.Stat_t).Uid)
+	gid = int(fi.Sys().(*syscall.Stat_t).Gid)
+
+	if uid > 0 && gid > 0 {
+		_ = os.Chown(dst, uid, gid)
+	}
 }
